@@ -17,22 +17,27 @@
   },
   "homepage": "https://github.com/pomo-mondreganto/CTForces#readme",
   "dependencies": {
-    "blueimp-file-upload-node": "^2.1.0",
+    "amaranjs": "^0.5.5",
     "css-loader": "^0.28.11",
+    "easytimer.js": "^2.2.3",
     "file-loader": "^1.1.11",
+    "github-markdown-css": "^2.10.0",
     "highlight.js": "^9.12.0",
     "jquery": "^3.3.1",
-    "jquery-ui": "^1.12.1",
-    "katex": "^0.10.0-beta",
-    "markdown-it": "^8.4.1",
+    "jquery-form": "^4.2.2",
+    "jquery-ui-dist": "^1.12.1",
+    "katex": "^0.10.0-rc.1",
+    "markdown-it": "^8.4.2",
     "markdown-it-emoji": "^1.4.0",
+    "markdown-it-katex": "^2.0.3",
     "markdown-it-sub": "^1.0.0",
     "markdown-it-sup": "^1.0.0",
-    "semantic-ui": "^2.3.2",
+    "semantic-ui": "^2.4.0",
     "semantic-ui-calendar": "0.0.8",
     "simplemde": "^1.11.2",
     "style-loader": "^0.21.0",
-    "webpack": "^4.14.0"
+    "tag-it": "^2.0.0",
+    "webpack": "^4.20.0"
   },
   "devDependencies": {}
 }
