@@ -1,17 +1,16 @@
 from django.conf import settings
-from django.contrib import messages
 from django.db.models import IntegerField, BooleanField
 from django.db.models import Q, Sum, Case, When, Value as V, Prefetch, Count, Subquery, OuterRef
 from django.http import Http404
 from django.http import JsonResponse
-from django.shortcuts import redirect, reverse
+from django.shortcuts import reverse
 from django.views.decorators.http import require_GET, require_POST
 from django.views.generic import TemplateView
 from guardian.shortcuts import get_objects_for_user, assign_perm
 
 from website.decorators import custom_login_required as login_required
 from website.forms import ContestForm
-from website.mixins import PermissionsRequiredMixin
+from website.mixins import AjaxPermissionsRequiredMixin
 from website.models import User, Contest, Task, TaskTag, ContestTaskRelationship
 from .view_classes import PagedTemplateView, UsernamePagedTemplateView, GetPostTemplateViewWithAjax
 
@@ -56,15 +55,16 @@ def submit_contest_flag(request, contest_id, task_id):
     response_dict = dict()
     if flag == task.flag:
         response_dict['success'] = True
-        if not task.solved_by.filter(id=request.user.id).exists() and not request.user.has_perm('edit_task', task):
+        if not request.user.has_perm('change_task', task):
             if contest.is_running:
                 relationship = task.contest_task_relationship.filter(contest=contest).first()
 
                 if not relationship:
                     raise Http404()
-
-                relationship.solved.add(request.user)
-            task.solved_by.add(request.user)
+                if not relationship.solved.filter(id=request.user.id).exists():
+                    relationship.solved.add(request.user)
+            if not task.solved_by.filter(id=request.user.id).exists():
+                task.solved_by.add(request.user)
 
         response_dict['next'] = reverse('contest_view', kwargs={'contest_id': contest_id})
     else:
@@ -85,22 +85,29 @@ def register_for_contest(request, contest_id):
     if not contest:
         raise Http404()
 
-    if not contest.is_registration_open:
-        return redirect('contests_main_list_view')
+    result = dict()
+
+    if not contest.is_registration_open or request.user.has_perm('change_contest', contest):
+        result['success'] = False
+        result['next'] = reverse('contests_main_list_view')
+        return JsonResponse(result)
 
     if contest.is_running or not contest.is_finished:
         if not contest.participants.filter(id=request.user.id).exists():
             contest.participants.add(request.user)
-            assign_perm('view_running_contest', request.user, contest)
+            assign_perm('can_participate_in_contest', request.user, contest)
 
+    result['success'] = True
     if contest.is_running:
-        return redirect('contest_view', contest_id=contest_id)
+        result['next'] = reverse('contest_view', kwargs=dict(contest_id=contest_id))
     else:
-        return redirect('contests_main_list_view')
+        result['next'] = reverse('contests_main_list_view')
+
+    return JsonResponse(result)
 
 
 class ContestMainView(TemplateView):
-    template_name = 'contest_view.html'
+    template_name = 'contest_templates/contest_view.html'
 
     def get_context_data(self, **kwargs):
         context = super(ContestMainView, self).get_context_data(**kwargs)
@@ -116,8 +123,6 @@ class ContestMainView(TemplateView):
         ).first()
         if not contest:
             raise Http404()
-        if contest.is_running and not self.request.user.has_perm('view_running_contest', contest):
-            return redirect('contests_main_list_view')
 
         tasks = contest.tasks.annotate(
             solved_count=Count(
@@ -127,7 +132,7 @@ class ContestMainView(TemplateView):
             is_solved_by_user=Sum(
                 Case(
                     When(
-                        solved_by__id=(self.request.user.id or -1),
+                        contest_task_relationship__solved__id=(self.request.user.id or -1),
                         then=1
                     ),
                     default=V(0),
@@ -141,14 +146,13 @@ class ContestMainView(TemplateView):
                 ).values('cost')
             )
         ).all()
-
         context['contest'] = contest
         context['tasks'] = tasks
         return context
 
 
 class ContestScoreboardView(PagedTemplateView):
-    template_name = 'contest_scoreboard.html'
+    template_name = 'contest_templates/contest_scoreboard.html'
 
     def get_context_data(self, **kwargs):
         context = super(ContestScoreboardView, self).get_context_data(**kwargs)
@@ -165,9 +169,7 @@ class ContestScoreboardView(PagedTemplateView):
         ).first()
 
         if not contest:
-            raise Http404()
-        if contest.is_running and not self.request.user.has_perm('view_running_contest', contest):
-            return redirect('contests_main_list_view')
+            raise Http404
 
         users = contest.participants.prefetch_related(
             Prefetch(
@@ -199,7 +201,7 @@ class ContestScoreboardView(PagedTemplateView):
 
 
 class ContestsMainListView(PagedTemplateView):
-    template_name = 'main_contests_list_view.html'
+    template_name = 'index_templates/main_contests_list_view.html'
 
     def get_context_data(self, **kwargs):
         context = super(ContestsMainListView, self).get_context_data(**kwargs)
@@ -223,7 +225,7 @@ class ContestsMainListView(PagedTemplateView):
 
 
 class UserContestListView(UsernamePagedTemplateView):
-    template_name = 'user_contests.html'
+    template_name = 'profile_templates/user_contests.html'
 
     def get_context_data(self, **kwargs):
         context = super(UserContestListView, self).get_context_data(**kwargs)
@@ -251,41 +253,43 @@ class UserContestListView(UsernamePagedTemplateView):
         return context
 
 
-class ContestCreationView(PermissionsRequiredMixin, GetPostTemplateViewWithAjax):
+class ContestCreationView(AjaxPermissionsRequiredMixin, GetPostTemplateViewWithAjax):
     permissions_required = (
         'add_contest',
     )
 
-    template_name = 'create_contest.html'
+    template_name = 'contest_templates/create_contest.html'
 
-    def handle_default(self, request, *args, **kwargs):
-
+    def handle_ajax(self, request, *args, **kwargs):
         task_ids = request.POST.getlist('tasks')
         task_tags = request.POST.getlist('tags')
         task_costs = request.POST.getlist('costs')
 
+        result = dict()
+
         if len(task_ids) != len(task_tags) or len(task_ids) != len(task_costs) or not len(task_ids):
-            messages.error(request=request, message='Need to add at least 1 task to contest.', extra_tags='task_id')
-            return redirect('create_contest')
+            result['success'] = False
+            result['errors']['tasks'] = 'Need to add at least 1 task to contest.'
+            result['next'] = reverse('create_contest')
+            return JsonResponse(result)
 
         tasks = []
         for i, task_id in enumerate(task_ids):
             task = Task.objects.filter(id=task_id).first()
-            if not task:
-                messages.error(request=request, message='Invalid task', extra_tags='task_id')
-                return redirect('create_contest')
 
-            if not request.user.has_perm('view_task', task):
-                messages.error(request=request, message='Invalid task', extra_tags='task_id')
-                return redirect('create_contest')
+            if not task or not request.user.has_perm('view_task', task):
+                result['success'] = False
+                result['errors']['tasks'] = 'Invalid task.'
+                result['next'] = reverse('create_contest')
+                return JsonResponse(result)
 
             tag_name = task_tags[i]
             tag = TaskTag.objects.filter(name=tag_name).first()
             if not tag:
-                messages.error(request=request,
-                               message='Invalid tag, only existing tags are allowed',
-                               extra_tags='task_tag')
-                return redirect('create_contest')
+                result['success'] = False
+                result['errors']['tags'] = 'Invalid tag, only existing tags are allowed.'
+                result['next'] = reverse('create_contest')
+                return JsonResponse(result)
 
             cost = task_costs[i]
             try:
@@ -293,10 +297,10 @@ class ContestCreationView(PermissionsRequiredMixin, GetPostTemplateViewWithAjax)
                 if cost < 0 or cost > 9999:
                     raise ValueError()
             except ValueError:
-                messages.error(request=request,
-                               message='Invalid cost. Cost must be an integer from 0 to 9999',
-                               extra_tags='task_cost')
-                return redirect('create_contest')
+                result['success'] = False
+                result['errors']['costs'] = 'Invalid cost. Cost must be an integer from 0 to 9999'
+                result['next'] = reverse('create_contest')
+                return JsonResponse(result)
 
             tasks.append((task, tag, cost))
 
@@ -310,21 +314,24 @@ class ContestCreationView(PermissionsRequiredMixin, GetPostTemplateViewWithAjax)
                                                        cost=task[2])
                 relationship.save()
 
+            assign_perm('change_contest', request.user, contest)
             assign_perm('view_unstarted_contest', request.user, contest)
-            assign_perm('view_running_contest', request.user, contest)
+            assign_perm('can_participate_in_contest', request.user, contest)
 
-            messages.success(request=request, message='Contest created successfully.')
-            return redirect('contest_view', contest_id=contest.id)
+            result['next'] = reverse('contest_view', kwargs=dict(contest_id=contest.id))
+            result['success'] = True
+            return JsonResponse(result)
         else:
             print(form.errors)
-            for field in form.errors:
-                for error in form.errors[field]:
-                    messages.error(request, error, extra_tags=field)
-            return redirect('create_contest')
+
+            result['next'] = reverse('create_contest')
+            result['success'] = False
+            result['errors'] = form.errors
+            return JsonResponse(result)
 
 
 class ContestTaskView(TemplateView):
-    template_name = 'contest_task.html'
+    template_name = 'task_templates/contest_task.html'
 
     def get_context_data(self, **kwargs):
         context = super(ContestTaskView, self).get_context_data(**kwargs)
@@ -339,9 +346,7 @@ class ContestTaskView(TemplateView):
             )
         ).first()
         if not contest:
-            raise Http404()
-        if contest.is_running and not self.request.user.has_perm('view_running_contest', contest):
-            return redirect('contests_main_list_view')
+            raise Http404
 
         task_id = kwargs.get('task_id')
         task = contest.tasks.filter(id=task_id).first()
